add validation

+++ simpleCurdApp/src/main/java/com/world/simpleCurdApp/Controllers/PersonController.java
@@ -3,6 +3,9 @@ package com.world.simpleCurdApp.Controllers;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+import javax.validation.ValidationException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.http.ResponseEntity;
@@ -20,11 +23,15 @@ public class PersonController {
 	private PersonReopsitory personReops;
 
 	@PostMapping(Routes.ADD_PERSON)
-	public String addpERSON(@RequestBody Person person) {
-		System.out.println(person);
+	public  ResponseEntity<String> addPerson(@Valid @RequestBody Person person) throws NotFoundException {
+		if(!personReops.findById(person.getId()).isPresent()){
 		personReops.save(person);
-		return "the Person hass been added";
-	}
+		return ResponseEntity.ok("User is valid");
+		}
+		else {
+			return ResponseEntity.ok("User is not valid or the user is already exist");
+		}
+		}
 
 	@GetMapping(Routes.getAllPeople)
 	public List<Person> getAll() {
@@ -37,13 +44,14 @@ public class PersonController {
 	}
 
 	@DeleteMapping(Routes.deletePersonById)
-	public String deletePerson(@PathVariable String id) {
-		try {
+	public ResponseEntity<String> deletePerson(@PathVariable String id) throws NotFoundException {
+		if(personReops.findById(id).isPresent()) {
 			personReops.deleteById(id);
-			return "the person hass been deleted";
-		} catch (Exception e) {
-			return e.toString();
-		}
+			return ResponseEntity.ok("the user hase been deleted");
+			}
+			else {
+				return ResponseEntity.ok("the user dose not found");	
+			}
 	}
 
 	@PutMapping(Routes.updatePersonById)
